+++ Configurateur/back/app.js
@@ -13,6 +13,7 @@ const feedbacksRoutes = require('./routes/Web/feedbacks');
 const itemsRoutes = require('./routes/Web/items');
 const paymentRoutes = require("./routes/Web/payment");
 const userMobileRoutes = require("./routes/Mobile/user");
+const authMobileRoutes = require("./routes/Mobile/auth");
 
 var cors = require("cors");
 var bodyParser = require("body-parser");
@@ -47,6 +48,7 @@ app.use('/api/messages', messagesRoutes)
 app.use('/api/feedbacks', feedbacksRoutes)
 app.use('/api/items', itemsRoutes)
 app.use("/api/payment", paymentRoutes);
-app.use("/api/mobile/user", userMobileRoutes);
+app.use("/api/mobile/user", userMobileRoutes)
+app.use("/api/mobile/auth", authMobileRoutes)
 
 module.exports = app;
