+++ Configurateur/back/controllers/Mobile/user.js
@@ -1,7 +1,5 @@
 const bcrypt = require('bcrypt')
-const jwt = require('jsonwebtoken')
 const { db } = require('../../middleware/database')
-const uuid = require('uuid')
 const transporter = require('../../middleware/transporter')
 
 /**
@@ -30,22 +28,6 @@ exports.findUserByEmail = email => {
       email: email
     }
   })
-}
-
-exports.sendAccountConfirmationEmail = (email, token) => {
-  let mailOptions = {
-    from: process.env.MAIL_ADDRESS,
-    to: email,
-    subject: 'Confirm your account',
-    text:
-      'Please follow the link to confirm your account: http://51.103.94.191:8080/api/mailVerification?token=' +
-      token
-  }
-  try {
-    transporter.sendMail(mailOptions)
-  } catch (error) {
-    console.error('Error sending reset password email:', error)
-  }
 }
 
 exports.getAllUsers = async () => {
@@ -104,17 +86,6 @@ exports.deleteUser = id => {
   return db.User_Mobile.delete({
     where: {
       id: id
-    }
-  })
-}
-
-exports.verifyEmail = id => {
-  return db.User_Mobile.update({
-    where: {
-      user: user.id
-    },
-    data: {
-      mailVerification: true
     }
   })
 }
